requirements.txt
asgiref==3.6.0
Django==4.1.5
django-autoslug==1.9.8
django-uuslug==2.0.0
djangorestframework==3.14.0
install==1.3.5
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
python-slugify==8.0.0
pytils==0.4.1
pytz==2022.7.1
sqlparse==0.4.3
text-unidecode==1.3
